new commit

+++ HomeWork7/Program.cs
@@ -1,7 +1,7 @@
 ﻿/*Показать вторую цифру трёхзначного числа*/
 
 Console.WriteLine("Введите трехзначное число ");
-double number = double.Parse(Console.ReadLine());
+int number = int.Parse(Console.ReadLine());
 
 string str_num = number.ToString();
 int a = int.Parse(str_num[0].ToString());
